chore(clerk-js): Refactor shared code
